two column chat and log
